Integrate Google Calendar Api into app
I can view calendars in the console, some methods
do not work. For example I can use 
user.calendar_client.list_calendar_lists,
but not user.calendar_client.calendars

I added the omniauth gem, and added the require
to the google api gem. This was supposed to be
'google/api_client' but it doesn't work.
access_type and scope were also added to the 
omniauth initializer. More info in link.

http://baugues.com/google-calendar-api-oauth2-and-ruby-on-rails

googleauth gem was added because the api's are
supposed to be authenticated with this library.