fix login for users who don't click links
I added a default role to the User model, to
fix the role is nil error. I also changed the 
user show page to reflect the role, and give
users the ability to change roles if they so 
desire.

I also cleaned up the layouts and added the 
login_path to routes.